# mobile/openapi.yaml
openapi: 3.0.3
info:
  title: ViVu App API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.vivuapp.vn
    description: Production
...
paths:
...
  /api/v1/posts/create:
    post:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Create post with files
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [postType, visibility]
              properties:
                content:
                  type: string
                postType:
                  type: string
                  enum: [EXPERIENCE, ITINERARY, GROUP_INVITE, TOUR, TOUR_SHARE]
                visibility:
                  type: string
                  enum: [PUBLIC, FRIENDS, PRIVATE]
                tourId:
                  type: string
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePostResponse"
  /api/v1/posts/{postId}:
    get:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Get post by id
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePostResponse"
  /api/v1/posts/{postId}/comments:
    get:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Get comments by post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Slice of comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCommentSlice"
  /api/v1/posts:
    get:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Get feed
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePostSlice"
  /api/v1/posts/me:
    get:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Get my posts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: My posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePostSlice"
  /api/v1/posts/{postId}/save:
    post:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Toggle save post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Saved/unsaved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseSavedPostResponse"
  /api/v1/posts/{postId}/share:
    post:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Share post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharePostRequest"
      responses:
        "200":
          description: Shared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseSharePostResponse"
  /api/v1/posts/me/saved-posts:
    get:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Get saved posts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Saved posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePostSlice"
  /api/v1/posts/{post_id}:
    delete:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Delete post
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
    put:
      tags: [Posts]
      security:
        - bearerAuth: []
      summary: Update post with files
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [postType, visibility]
              properties:
                content:
                  type: string
                postType:
                  type: string
                  enum: [EXPERIENCE, ITINERARY, GROUP_INVITE, TOUR, TOUR_SHARE]
                visibility:
                  type: string
                  enum: [PUBLIC, FRIENDS, PRIVATE]
                tourId:
                  type: string
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePostResponse"
...
  /api/v1/likes/post/{postId}:
    post:
      tags: [Reactions]
      security:
        - bearerAuth: []
      summary: Toggle like on post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Liked/unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLikeResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        result:
          nullable: true
    ApiResponseVoid:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
...
    ApiResponsePostResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/PostResponse"
    ApiResponsePostSlice:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/SlicePostResponse"
...
    ApiResponseCommentSlice:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/SliceCommentResponse"
    ApiResponseSavedPostResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/SavedPostResponse"
    ApiResponseSharePostResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/SharePostResponse"
    ApiResponseLikeResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/LikeResponse"
...
    SharePostRequest:
      type: object
      required: [conversationId]
      properties:
        conversationId:
          type: string
        message:
          type: string
    PostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        postType:
          $ref: "#/components/schemas/PostType"
        visibility:
          $ref: "#/components/schemas/Visibility"
        authorId:
          type: integer
          format: int64
        authorName:
          type: string
        authorAvatarUrl:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        authorId:
          type: integer
          format: int64
        authorName:
          type: string
        authorAvatarUrl:
          type: string
        postId:
          type: integer
          format: int64
        parentCommentId:
          type: integer
          format: int64
        likeCount:
          type: integer
        timeStamp:
          type: string
          format: date-time
    SavedPostResponse:
      type: object
      properties:
        postId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        isSaved:
          type: boolean
    SharePostResponse:
      type: object
      properties:
        messageId:
          type: integer
          format: int64
        postId:
          type: integer
          format: int64
        conversationId:
          type: string
        senderId:
          type: integer
          format: int64
        senderName:
          type: string
        message:
          type: string
        messageType:
          $ref: "#/components/schemas/MessageType"
        timestamp:
          type: string
          format: date-time
    LikeResponse:
      type: object
      properties:
        postId:
          type: integer
          format: int64
        commentId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        isLiked:
          type: boolean
        likeCount:
          type: integer
          format: int64
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        offset:
          type: integer
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    SlicePostResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PostResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        size:
          type: integer
        number:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        numberOfElements:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    SliceCommentResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CommentResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        size:
          type: integer
        number:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        numberOfElements:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
    PostType:
      type: string
      enum: [EXPERIENCE, ITINERARY, GROUP_INVITE, TOUR, TOUR_SHARE]
    Visibility:
      type: string
      enum: [PUBLIC, FRIENDS, PRIVATE]
...
    MessageType:
      type: string
      enum: [TEXT, IMAGE, VIDEO, AUDIO, DOCUMENT, SHARED_POST]
...